+++ ui-frameworks/shadcn/js/src/index.jsx
@@ -32,12 +32,16 @@ import { Toaster } from "@/components/sonner";
 import { Toggle } from "@/components/toggle";
 import { Avatar } from "@/components/avatar";
 import { AlertDialog } from "@/components/alert-dialog";
+import { Carousel } from "@/components/carousel";
+import { Chart } from "@/components/chart";
 import { Command } from "@/components/command";
 import { ContextMenu } from "@/components/context-menu";
 import { Drawer } from "@/components/drawer";
+import { InputOtp } from "@/components/input-otp";
 import { HoverCard } from "@/components/hover-card";
 import { Menubar } from "@/components/menubar";
 import { NavigationMenu } from "@/components/navigation-menu";
+import { Resizable } from "@/components/resizable";
 import { ScrollArea } from "@/components/scroll-area";
 import { Sheet } from "@/components/sheet";
 import { Tooltip } from "@/components/tooltip";
@@ -75,12 +79,16 @@ window.shinyreact.registerComponents(null, {
   "shadcn:Toggle": Toggle,
   "shadcn:Avatar": Avatar,
   "shadcn:AlertDialog": AlertDialog,
+  "shadcn:Carousel": Carousel,
+  "shadcn:Chart": Chart,
   "shadcn:Command": Command,
   "shadcn:ContextMenu": ContextMenu,
   "shadcn:Drawer": Drawer,
   "shadcn:HoverCard": HoverCard,
   "shadcn:Menubar": Menubar,
+  "shadcn:InputOtp": InputOtp,
   "shadcn:NavigationMenu": NavigationMenu,
+  "shadcn:Resizable": Resizable,
   "shadcn:ScrollArea": ScrollArea,
   "shadcn:Sheet": Sheet,
   "shadcn:Tooltip": Tooltip,